refactor: Limited the amount of active tasks queued at the same time in ModUpdateAvailableChecker.cs. This should improve performance when checking for mod updates with a large amount of mods.
Before this change, the ModUpdateAvailableChecker would queue a new task for each mod that needed to be checked for updates.
With a lot of mods, a large amount of tasks would be queued at the same time, because of the rate limiter they would all be rejected and the tasks would be queued again.
This change limits the amount of active tasks that can be queued at the same time to 20, this should improve app responsiveness during a mod update check, especially with a large amount of mods.