refactor: Limited the number of active tasks queued at the same time in ModUpdateAvailableChecker. This should improve performance when checking for mod updates with a large number of mods. (#214)
Before this change, the ModUpdateAvailableChecker would queue a new task for each mod that needed to be checked for updates.
With a lot of mods, a large number of tasks would be queued at the same time, because of the rate limiter they would all be rejected, and the tasks would be requeued again. With many mods the UI would become laggy.
This change limits the number of active tasks that can be queued at the same time to 20, this should improve app responsiveness during a mod update check, especially with a large number of mods.